Setting up authentication on frontend.

+++ frontend/src/components/WorkoutDetails.js
@@ -1,11 +1,19 @@
 import { useWorkoutContext } from "../hooks/useWorkoutContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 
 const WorkoutDetails = ({workout}) => {
     const {dispatch} = useWorkoutContext();
+    const {user} = useAuthContext();
     const handleClick = async () => {
+        if (!user) {
+            return;
+        }
         const response = await fetch(`${process.env.REACT_APP_API_URL}/api/workouts/${workout._id}`, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
 
         const json = await response.json();
